+++ embabel-modules/embabel-streaming/README.md
@@ -7,18 +7,13 @@ Integration tests for Embabel Agent streaming with OpenAI and Anthropic.
 Export your API keys before running:
 
 ```bash
-export OPENAI_API_KEY=your-openai-key
 export ANTHROPIC_API_KEY=your-anthropic-key
 ```
 
 ## Run
 
-**OpenAI tests** (gpt-4.1-mini):
-```bash
-mvn test -pl embabel-streaming-openai
-```
-
 **Anthropic tests** (claude-sonnet-4-5):
 ```bash
-mvn test -pl embabel-streaming-anthropic
+mvn test -pl embabel-streaming -P integration
+
 ```
